requirements.txt
numpy
sentence-transformers
scikit-learn
google-genai
torch
streamlit